+++ bare-metal-apps/arch/riscv/qemu-virt/platform.c
@@ -88,7 +88,7 @@ void init_proc(){
  * @param ms_delay Time period.
  * @return 0 if the period was not consumed, 1 otherwise.
  */
-uint32_t wait_time(uint32_t old_time, uint32_t ms_delay){
+uint32_t wait_time(uint64_t old_time, uint64_t ms_delay){
 	uint64_t diff_time;
 	uint64_t now = get_mtimer_value();
